refactor: Recommendation API 구조 개선
