+++ src/main/java/swyp12/team9/server/api/recommendation/dto/RecommendationResponse.java
@@ -1,4 +1,4 @@
-package swyp12.team9.server.api.dto;
+package swyp12.team9.server.api.recommendation.dto;
 
 public record RecommendationResponse(
     Long id,
