+++ src/main/java/swyp12/team9/server/api/recommendation/dto/SimilarContentResponse.java
@@ -1,4 +1,4 @@
-package swyp12.team9.server.api.dto;
+package swyp12.team9.server.api.recommendation.dto;
 
 public record SimilarContentResponse(
     RecommendationResponse content,
